drafted temporal adj for non-hwy with reduced windows & pcts

+++ R - hei_aim3a/r0_temporal_adjustment.R
@@ -48,9 +48,6 @@ message("loading data")
 fixed_site_temp_adj <- readRDS(file.path("data", "epa_data_mart", "wa_county_nox_temp_adjustment.rda")) %>%
   select(time, ufp_adjustment = diff_adjustment_winsorize)
 
-# all visit medians
-pnc_meds <- readRDS(file.path(dt_pt, "pnc_med.rds"))
-
 # BH samples
 if(!file.exists(file.path(dt_out, "TEMP_bh_visits.rda"))) {
   v1 <- readRDS(file.path(dt_pt, "nonspatial_visit_samples.rds")) %>%
@@ -62,10 +59,10 @@ if(!file.exists(file.path(dt_out, "TEMP_bh_visits.rda"))) {
   visits <- bind_rows(v1, v2) %>%
     select(-c(visit_num, runname)) 
   
-  saveRDS(visits, file.path(dt_out, "TEMP_bh_visits.rda"))
+  saveRDS(visits, file.path(dt_pt, "TEMP_bh_visits.rda"))
   rm(list=c("v1", "v2"))
 } else {
-    visits <- readRDS(file.path(dt_out, "TEMP_bh_visits.rda"))
+    visits <- readRDS(file.path(dt_pt, "TEMP_bh_visits.rda"))
   }
 
 
@@ -74,7 +71,7 @@ if(!file.exists(file.path(dt_out, "TEMP_bh_visits.rda"))) {
 ## note some measurements are repeated for same time, e.g. 2019-02-25 18:55:23 for I405 and I5 (UFP is the same)
 ## n= 3,688,353 (not what was reported by Annie: 3,769,325 1s measures)
 
-if(!file.exists(file.path(dt_pt, "TEMP_road_dt.rda"))) {
+if(!file.exists(file.path(dt_pt, "TEMP_road_dt.rda") | !file.exists(file.path(dt_pt, "TEMP_road_dt_no_hwy.rda")))) {
   road_dt0 <- readRDS(file.path("data", "onroad", "annie", "OnRoad Paper Code Data", "data", "All_Onroad_12.20.rds")) %>%
     arrange(time) %>%
     mutate(
@@ -96,12 +93,14 @@ if(!file.exists(file.path(dt_pt, "TEMP_road_dt.rda"))) {
   
   saveRDS(road_dt0, file.path(dt_pt, "TEMP_road_dt0.rda"))
   
-  # apply(road_dt0, 2, function(x) any(is.na(x)))
-  
   missingness_table <- count_missingness(road_dt0, notes="only keep 1st location when same sec reading is assigned to 2")
+  
+  # drop hwy for some of these
+  road_dt0_no_hwy <- filter(road_dt0, road_type != "A1")
+  missingness_table <- count_missingness(road_dt0_no_hwy, notes="drop A1 segments")
+  
   write.csv(missingness_table, file.path(dt_pt, "missing_counts_second_dt.csv"), row.names = F)
   
-  # x=unique(road_dt$runname)[1]
   time_series <- lapply(unique(road_dt0$runname), function(x){
     a_run <- filter(road_dt0, runname==x)
     data.frame(runname = x,
@@ -111,18 +110,18 @@ if(!file.exists(file.path(dt_pt, "TEMP_road_dt.rda"))) {
   
   # NAs in hour start here b/c of missing readings
   road_dt <- left_join(time_series, road_dt0, by=c("runname", "time")) %>%
-    select(runname, time, hour, weekday, id, ufp) 
+    select(runname, time, hour, weekday, id, ufp)
   
-  # apply(road_dt, 2, function(x) sum(is.na(x)))
-  
-  rm(time_series)
+  road_dt_no_hwy <- left_join(time_series, road_dt0_no_hwy, by=c("runname", "time")) %>%
+    select(runname, time, hour, weekday, id, ufp)
   
   saveRDS(road_dt, file.path(dt_pt, "TEMP_road_dt.rda"))
+  saveRDS(road_dt_no_hwy, file.path(dt_pt, "TEMP_road_dt_no_hwy.rda"))
   
-  # 25% of second data are missing here (e.g., from stops)
-  # (nrow(time_series) - nrow(road_dt0))/nrow(time_series)
+  rm(list=c("time_series", "road_dt0", "road_dt0_no_hwy"))
 } else {
   road_dt <- readRDS(file.path(dt_pt, "TEMP_road_dt.rda"))
+  road_dt_no_hwy <- readRDS(file.path(dt_pt, "TEMP_road_dt_no_hwy.rda"))
   }
 
 
@@ -157,34 +156,74 @@ if(!file.exists(file.path(dt_pt, "TEMP_road_dt.rda"))) {
 ##################################################################################################
 # VARIABLES
 ##################################################################################################
-business_hours <- c(9:17)
+# e.g., 3 hr * 60 min/hr * 60 sec/min
+# windows <- 60*60#c(1:3)*60*60  
+# quantiles <- 0.10 #c(0.01, 0.05, 0.10, 0.20)
+windows <- c(1,3)*60*60
+# --> could drop 0.20
+quantiles <- c(0.01, 0.05#, 0.10#, 0.20
+               )
+
+##################################################################################################
+# FUNCTIONS
+##################################################################################################
+
+calculate_rolling_quantile <- function(dt, windows.=windows, quantiles.=quantiles) {
+  
+  adj_lvls <- apply(expand.grid(paste0("hr", sort(windows)/3600), 
+                                paste0("_pct", sort(quantiles)*100)), 
+                    1, paste, collapse="")
+  
+  dt <- lapply(windows., function(w) {
+    lapply(quantiles., function(p) {
+      
+      bg_label <- paste0("hr", w/3600, "_pct", p*100)
+      
+      dt %>%
+        group_by(runname) %>%
+        mutate(
+          background_adj = bg_label,
+          # rolling quantile, ignore NAs (i.e., periods of time w/o UFP readings) may still get background estimates
+          background = rollapply(ufp, width = w, FUN = quantile, prob = p, align = "center", partial = TRUE, na.rm=T))
+    })
+  }) %>%
+    bind_rows() %>%
+    mutate(background_adj = factor(background_adj, levels=adj_lvls))
+}
 
 ##################################################################################################
 # 1. TEMPORAL ADJUSTMENT: PSEUDO FIXED SITES (FROM PREDICTED UFP)
 ##################################################################################################
-message("running fixed site temporal adjustment from predicted UFP based on NO2")
 
-visits_adj1 <- visits %>%
-  mutate(time = ymd_h(paste(date, hour))) %>%
-  # add temporal adjustment
-  left_join(fixed_site_temp_adj, by="time") %>%
-  mutate(median_value_adjusted = median_value + ufp_adjustment,
-         version = paste(version, "temp adj 1"))
+if(!file.exists(file.path(dt_pt, "TEMP_bh_site_avgs_fixed_site_temporal_adj.rds"))) {
+  message("running fixed site temporal adjustment from predicted UFP based on NO2")
 
-saveRDS(visits_adj1, file.path(dt_pt, "bh_visits_fixed_site_temporal_adj.rds")) 
+  visits_adj1 <- visits %>%
+    mutate(time = ymd_h(paste(date, hour))) %>%
+    # add temporal adjustment
+    left_join(fixed_site_temp_adj, by="time") %>%
+    mutate(median_value_adjusted = median_value + ufp_adjustment,
+           version = paste(version, "temp adj 1"))
 
-annual_adj1 <- visits_adj1 %>%
-  group_by(id, adjusted, actual_visits, campaign, design, visits, version, cluster_type, cluster_value) %>%
-  summarize(annual_mean = mean(median_value_adjusted, na.rm=T)) %>%
-  ungroup()  
+  saveRDS(visits_adj1, file.path(dt_pt, "bh_visits_fixed_site_temporal_adj.rds"))
 
-# temp file
-saveRDS(annual_adj1, file.path(dt_pt, "TEMP_bh_site_avgs_fixed_site_temporal_adj.rds"))
+  annual_adj1 <- visits_adj1 %>%
+    group_by(id, adjusted, actual_visits, campaign, design, visits, version, cluster_type, cluster_value) %>%
+    summarize(annual_mean = mean(median_value_adjusted, na.rm=T)) %>%
+    ungroup()
+
+  # temp file
+  saveRDS(annual_adj1, file.path(dt_pt, "TEMP_bh_site_avgs_fixed_site_temporal_adj.rds"))
+} else {
+  annual_adj1 <- readRDS(file.path(dt_pt, "TEMP_bh_site_avgs_fixed_site_temporal_adj.rds"))
+}
 
 ##################################################################################################
-# QC - check that few annual averages are <=0
-summary(annual_adj1$annual_mean)
-prop.table(table(annual_adj1$annual_mean<=0))
+# # QC - check that few annual averages are <=0
+# message("temp adj 1 - predicted UFP from NO2")
+# summary(annual_adj1$annual_mean)
+# table(annual_adj1$annual_mean<=0)
+# prop.table(table(annual_adj1$annual_mean<=0))
 
 
 ##################################################################################################
@@ -193,90 +232,44 @@ prop.table(table(annual_adj1$annual_mean<=0))
 message("running fixed site temporal adjustment from the collected UFP measures")
 
 if(!file.exists(file.path(dt_pt, "underwrite_temp_adj_all_1s_data.rda"))) {
-  # e.g., 3 hr * 60 min/hr * 60 sec/min
-  
-  # windows <- 60*60#c(1:3)*60*60  
-  # quantiles <- 0.10 #c(0.01, 0.05, 0.10, 0.20)
-  windows <- c(1:3)*60*60
-  quantiles <- c(0.01, 0.05, 0.10, 0.20)
-  
-  road_dt <- lapply(windows, function(w) {
-    lapply(quantiles, function(p) {
-      
-      bg_label <- paste0("hr", w/3600, "_pct", p*100)
-      
-      road_dt %>%
-        group_by(runname) %>%
-        mutate(
-          background_adj = bg_label,
-          # rolling quantile, ignore NAs (i.e., periods of time w/o UFP readings) may still get background estimates
-          background = rollapply(ufp, width = w, FUN = quantile, prob = p, align = "center", partial = TRUE, na.rm=T))
-    })
-  }) %>%
-    bind_rows()
-  
-  # --> COULD ADD OVERNIGHT ADJUSTMENT SINCE THERE IS NO MM DATA DURING THESE TIMES
-  
+  road_dt <- calculate_rolling_quantile(dt=road_dt)
   saveRDS(road_dt, file.path(dt_pt, "underwrite_temp_adj_all_1s_data.rda"))
-} else {
-  road_dt <- readRDS(file.path(dt_pt, "underwrite_temp_adj_all_1s_data.rda"))
-  }
+  } else {
+    road_dt <- readRDS(file.path(dt_pt, "underwrite_temp_adj_all_1s_data.rda"))
+    }
 
-# note: some hours don't have UFP but still have hourly adjustments b/c rm.na=T for rolling window calculations
-underwrite_adj <- road_dt %>%
-  group_by(background_adj) %>%
-  mutate(bg_lta = mean(background, na.rm = T),
-         date= as.Date(time)) %>%
-  group_by(runname, date, hour, background_adj, bg_lta) %>%
-  summarize(bg_hour_avg = mean(background, na.rm = T),
-            bg_hour_median = median(background, na.rm = T)) %>%
-  ungroup() %>%
-  mutate(avg_hourly_adj = bg_lta - bg_hour_avg,
-         median_hourly_adj = bg_lta - bg_hour_median,
-         time = ymd_h(paste(date, hour))) %>%
-  select(runname, date, hour, time, background_adj, bg_lta, bg_hour_avg, avg_hourly_adj, bg_hour_median, median_hourly_adj)
-
-# apply(underwrite_adj, 2, function(x) sum(is.na(x)))
-
-saveRDS(underwrite_adj, file.path(dt_pt, "underwrite_temp_adj.rda"))
-
-# underwrite_adj <- readRDS(file.path(dt_pt, "underwrite_temp_adj.rda"))
-
+if(!file.exists(file.path(dt_pt, "underwrite_temp_adj_all_1s_data_no_hwy.rda"))) {
+  road_dt_no_hwy <- calculate_rolling_quantile(dt=road_dt_no_hwy)
+  saveRDS(road_dt_no_hwy, file.path(dt_pt, "underwrite_temp_adj_all_1s_data_no_hwy.rda"))
+  } else {
+    road_dt_no_hwy <- readRDS(file.path(dt_pt, "underwrite_temp_adj_all_1s_data_no_hwy.rda"))
+    }
 
 ##################################################################################################
-# EXAMPLE - visualize hourly adjustments
-adj_lvls <- apply(expand.grid(paste0("hr", c(1:3)), 
-                              paste0("_pct", c(1, 5, 10, 20))), 
-                  1, paste, collapse="")
-  
+# note: some hours don't have UFP but still have hourly adjustments b/c rm.na=T for rolling window calculations
+get_hourly_adjustment <- function(dt) {
+  dt %>%
+    group_by(background_adj) %>%
+    mutate(bg_lta = mean(background, na.rm = T),
+           date= as.Date(time)) %>%
+    group_by(runname, date, hour, background_adj, bg_lta) %>%
+    summarize(bg_hour_avg = mean(background, na.rm = T),
+              bg_hour_median = median(background, na.rm = T)) %>%
+    ungroup() %>%
+    mutate(avg_hourly_adj = bg_lta - bg_hour_avg,
+           median_hourly_adj = bg_lta - bg_hour_median,
+           time = ymd_h(paste(date, hour))) %>%
+    select(runname, date, hour, time, background_adj, bg_lta, bg_hour_avg, avg_hourly_adj, bg_hour_median, median_hourly_adj)
+  }
+##################################################################################################
 
-underwrite_adj %>%
-  drop_na(hour) %>%
-  mutate(background_adj = factor(background_adj, levels=adj_lvls)) %>%
-  # couple of day examples
-  filter(runname %in% unique(underwrite_adj$runname)[1:6]) %>%  
-  
-  ggplot(aes(x=hour, y=avg_hourly_adj, group=background_adj, col=background_adj)) + 
-  facet_wrap(facets = vars(runname)) +
-  geom_hline(yintercept = 0, linetype=2, alpha=0.5) +
-  geom_point() + 
-  geom_line() + 
-  scale_color_ordinal()
+underwrite_adj <- get_hourly_adjustment(road_dt)
+underwrite_adj_no_hwy <- get_hourly_adjustment(road_dt_no_hwy)
 
-ggsave(file.path(dt_out, "temp_adj2_example.png"), width = 10, height = 6)
+saveRDS(underwrite_adj, file.path(dt_pt, "underwrite_temp_adj.rda"))
+saveRDS(underwrite_adj_no_hwy, file.path(dt_pt, "underwrite_temp_adj_no_hwy.rda"))
 
-# time series
-underwrite_adj %>%
-  drop_na(hour) %>%
-  mutate(background_adj = factor(background_adj, levels=adj_lvls)) %>%
-  
-  ggplot(aes(x=time, y=avg_hourly_adj)) + 
-  facet_wrap(facets = vars(background_adj)) +
-  geom_hline(yintercept = 0, linetype=2, alpha=0.5) +
-  geom_point(alpha=0.4, aes(col=background_adj)) + 
-  geom_smooth()
-   
-ggsave(file.path(dt_out, "temp_adj2_time_series.png"), width = 14, height = 8)
+# underwrite_adj <- readRDS(file.path(dt_pt, "underwrite_temp_adj.rda"))
 
 ##################################################################################################
 visits_adj2 <- visits %>%
@@ -310,6 +303,102 @@ summary(annual_adj2$annual_mean)
 prop.table(table(annual_adj2$annual_mean<=0))
 
 
+
+
+##################################################################################################
+# QC - INVESTIGATE NEGATIVE ANNUAL AVERAGES
+##################################################################################################
+# --> COMPARE CONC VISITS VS ADJUSTMENTS
+
+
+
+
+
+
+##################################################################################################
+# QC - VISUALIZE TEMPORAL ADJUSTMENT PATTERNS 
+##################################################################################################
+# TIME SERIES: ONROAD MEASURES VS 'BACKGROUND' (E.G., Brantley Fig 5)
+
+# --> 
+
+
+
+##################################################################################################
+# --> plot percentiles: dots w/ smoother
+# --> add example of ~1 wk 
+# --> consider overlaying same route driven on same time…driving same structure
+
+
+
+
+##################################################################################################
+# # hourly adjustments
+# adj_lvls <- apply(expand.grid(paste0("hr", windows/3600), 
+#                               paste0("_pct", quantiles*100)), 
+#                   1, paste, collapse="")
+# 
+# 
+# underwrite_adj %>%
+#   drop_na(hour) %>%
+#   mutate(background_adj = factor(background_adj, levels=adj_lvls)) %>%
+#   # couple of day examples
+#   filter(runname %in% unique(underwrite_adj$runname)[1:6]) %>%  
+#   
+#   ggplot(aes(x=hour, y=avg_hourly_adj, group=background_adj, col=background_adj)) + 
+#   facet_wrap(facets = vars(runname)) +
+#   geom_hline(yintercept = 0, linetype=2, alpha=0.5) +
+#   geom_point() + 
+#   geom_line() + 
+#   scale_color_ordinal()
+# 
+# ggsave(file.path(dt_out, "temp_adj2_example.png"), width = 10, height = 6)
+# 
+# # time series
+# underwrite_adj %>%
+#   drop_na(hour) %>%
+#   mutate(background_adj = factor(background_adj, levels=adj_lvls)) %>%
+#   
+#   ggplot(aes(x=time, y=avg_hourly_adj)) + 
+#   facet_wrap(facets = vars(background_adj)) +
+#   geom_hline(yintercept = 0, linetype=2, alpha=0.5) +
+#   geom_point(alpha=0.4, aes(col=background_adj)) + 
+#   geom_smooth()
+# 
+# ggsave(file.path(dt_out, "temp_adj2_time_series.png"), width = 14, height = 8)
+# 
+
+
+
+##################################################################################################
+# QC - VISUALIZE SPATIAL ADJUSTMENT PATTERNS 
+##################################################################################################
+# maps to make sure background doesnt vary by space
+
+
+
+
+
+##################################################################################################
+# SAVE DATA
+################################################################################################
+# message("saving final files")
+# rbind(annual_adj1, annual_adj2) %>%
+#   saveRDS(., file.path(dt_pt, "temp_adj_site_avgs.rds"))
+
+##################################################################################################
+# DONE
+##################################################################################################
+message("DONE RUNNING R0_TEMPORAL_ADJUSTMENT.R")
+
+
+
+
+
+
+##################################################################################################
+# APPENDIX
+##################################################################################################
 
 # ##################################################################################################
 # # 2. BACKGROUND ADJUSTMENT USING THE COLLECTED DATA
@@ -364,26 +453,6 @@ prop.table(table(annual_adj2$annual_mean<=0))
 # # temp file
 # saveRDS(annual_adj2, file.path(dt_pt, "bh_site_avgs_background_temporal_adj.rds"))
 
-
- 
-
-
-
-
-
-
-
-##################################################################################################
-# SAVE DATA
-################################################################################################
-message("saving final files")
-rbind(annual_adj1, annual_adj2) %>%
-  saveRDS(., file.path(dt_pt, "temp_adj_site_avgs.rds"))
-
-##################################################################################################
-# DONE
-##################################################################################################
-message("DONE RUNNING R0_TEMPORAL_ADJUSTMENT.R")
